% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRmdEquation.R
\name{addRdEquation}
\alias{addRdEquation}
\alias{addRdEquation_withLine}
\title{Addin: add an equation to a Rd help file}
\usage{
addRdEquation()

addRdEquation_withLine()
}
\value{
Null
}
\description{
Two functions
\itemize{
\item addRdEquation() Gives \\eqn{}
\item addRdEquation_withLine() Gives \\deqn{}
}
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 addRdEquation()
 addRdEquation_withLine()
 }
}
}
\seealso{
\code{\link[rstudioapi]{rstudio-documents}}
}
