% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_renv.R
\name{check_renv}
\alias{check_renv}
\title{Check if renv is ok or do a restore and snapshot}
\usage{
check_renv(synchronized = FALSE)
}
\arguments{
\item{synchronized}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[renv]{status}}, \code{\link[renv]{restore}}, \code{\link[renv]{snapshot}}
}
