% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_template.R
\name{citation_template}
\alias{citation_template}
\title{Preview or create a citation template}
\usage{
citation_template(file = "citation_template.tmp", preview = TRUE)
}
\arguments{
\item{file}{character, Default: 'citation_template.tmp'}

\item{preview}{logical, Default: TRUE}
}
\value{
Formatted citation: character
}
\description{
Inspire de usethis::use_citation(). Voir aussi la fonction use_template() pour utiliser
differents templates (dont pour les options ou les yaml). Pour obtenir le chemin au fichier de
citation d'une librairie: fs::path_package(package = "usethis", "templates", "citation-template.R"). Et
pour lire ce fichier: temp <- readLines(fs::path_package(package = "usethis", "templates", "citation-template.R"))
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 citation_template()
 citation_template(preview=FALSE)
 }
}
}
\seealso{
\code{\link[fs]{file_access}}, \code{\link[fs]{create}}
\code{\link[rstudioapi]{getActiveProject}}
}
