% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipToRprofile.R
\name{clipToRprofile}
\alias{clipToRprofile}
\title{Copy Rprofile_template to clipboard and open .Rprofile to edit}
\usage{
clipToRprofile()
}
\value{
No return
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[fs]{file_access}}
\code{\link[utils]{clipboard}}
\code{\link[usethis]{edit_file}}
}
