% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_objects.R
\name{list_objects}
\alias{list_objects}
\title{Liste des objects disponibles dans une librairie}
\usage{
list_objects(package, pattern = "")
}
\arguments{
\item{package}{character, package name}

\item{pattern}{character, Default: ''}
}
\value{
liste, Liste des objects disponibles, sinon NULL
}
\description{
Liste des objects disponibles dans une librairie
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 list_objects("raichepackage")
 }
}
}
\seealso{
\code{\link[utils]{ls.str}}
}
