% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_build.R
\name{pack_build}
\alias{pack_build}
\title{Creation des versions source et binaire d'une librairie}
\usage{
pack_build()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[devtools]{build}}
}
