% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_develop.R
\name{pack_develop}
\alias{pack_develop}
\title{Developpement d'une librairie}
\usage{
pack_develop()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[devtools]{load_all}}, \code{\link[devtools]{document}}, \code{\link[devtools]{build_rmd}}
}
