% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_initialisation.R
\name{pack_initialisation}
\alias{pack_initialisation}
\title{Initialisation des fichiers et repertoires d'une librarie}
\usage{
pack_initialisation(package)
}
\arguments{
\item{package}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[devtools]{reexports}}
\code{\link[usethis]{use_description}}, \code{\link[usethis]{use_author}}, \code{\link[usethis]{licenses}}, \code{\link[usethis]{use_roxygen_md}}, \code{\link[usethis]{use_citation}}, \code{\link[usethis]{use_addin}}, \code{\link[usethis]{use_vignette}}, \code{\link[usethis]{use_news_md}}, \code{\link[usethis]{use_readme_rmd}}, \code{\link[usethis]{use_logo}}
}
