% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_vignette.R
\name{pack_vignette}
\alias{pack_vignette}
\title{FCreate a vignette template and a vignette directory}
\usage{
pack_vignette(copyTo = "introduction", overwrite = FALSE)
}
\arguments{
\item{copyTo}{character, Default: 'template_html_Rmd'}

\item{overwrite}{logical, Default: FALSE}
}
\value{
No return
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[fs]{file_access}}, \code{\link[fs]{create}}, \code{\link[fs]{copy}}
}
