% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageInfo.R
\name{packageInfo}
\alias{packageInfo}
\title{Informations sur une librairie disponible sur CRAN ou Github}
\usage{
packageInfo(pkg)
}
\arguments{
\item{pkg}{character: package name}
}
\value{
Path to the package: character
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 repos = getOption("repos")
 packageInfo("ggplot2")
 }
}
}
\seealso{
\code{\link[pkgsearch]{cran_package}}
\code{\link[utils]{packageDescription}}
}
